penyedia-jasa-pembayaran/services/fraud-detection-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
grpcio==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.4
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.15.2
redis[hiredis]==5.3.0
pydantic-settings==2.8.1
pydantic==2.11.1
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-instrumentation-logging==0.49b2
opentelemetry-instrumentation-sqlalchemy==0.49b2
opentelemetry-instrumentation-redis==0.49b2
python-logging-loki==0.3.1
psycopg2-binary==2.9.10